First cut at "new fields" for atmosphere dump files
Add new data fields to an atmosphere dump file based on cloning
existing fields then replacing their data with values drawn from a
NetCDF file; also replace the stashcode and PP code for the new field
based on namelist input.

This turns out to be a tiny bit tricky, mostly because of managing
sizes and file offsets in the UM dump file format.  New fields are
inserted so as to preserve stashcode ordering of data fields in the
existing file: i.e., if the input dump file has data ordered in
stashcode order, then so will the output file.

The logic for "replacing" the existing field data with values from
NetCDF files is identical to the existing functionality for pulling
data for existing fields from NetCDF files.  The code works by
massaging the dump file lookup table to make space for the new fields,
the just running through all fields as normal, drawing the data to be
written to the output file either from the original dump file or from
a NetCDF file.  This means that the error handling for newly inserted
fields is identical to that for modification of existing fields.