Change date handling for dump files
It seems as though the fixhd(21:27) and fixhd(28:34) fields in the
headers of dump files generally hold the start and end dates of the
simulation that generated the dump file.  xconv relies on these being
the appropriate distance apart in time.  I've fixed this up a bit, but
xconv will still crash if you attempt to change anything but the year
of the reference date.