Fix fixed header array content detection; add level code for ocean
I've now seen dump files that show empty arrays (real constants, etc.)
in the fixed header by setting the offset to -32768, instead of
setting the length to zero, so I've changed the way this is detected.

Also, I've now seen an ocean dump file with level code = 1 for some
variables, so I've added that as a (single level) valid level code for
the ocean.